Revert "EVG-19211 migrate to v2 of the AWS go SDK (#6579)"
This reverts commit 7f39e06a3115986934a6bd1933e172a5049bb859.

+++ operations/host_spawn.go
@@ -829,7 +829,7 @@ func hostModifyVolume() cli.Command {
 
 			opts := restModel.VolumeModifyOptions{
 				NewName:       name,
-				Size:          int32(size),
+				Size:          size,
 				NoExpiration:  noExpiration,
 				HasExpiration: hasExpiration,
 			}
@@ -952,7 +952,7 @@ func hostCreateVolume() cli.Command {
 
 			volumeRequest := &host.Volume{
 				Type:             volumeType,
-				Size:             int32(volumeSize),
+				Size:             volumeSize,
 				AvailabilityZone: volumeZone,
 				DisplayName:      volumeName,
 			}
